requirements.txt
tensorflow==2.10.0
pyspark==3.3.0
pyarrow==9.0.0
boto3==1.24.96
Pillow==9.2.0
jupyter==1.0.0
jupyterlab==3.4.8